feat: complete Rust for Linux kernel integration
Major changes:
- Add kernel module with FFI bindings (src/kernel/)
- Fix all clippy warnings (0 warnings with --features kernel)
- Refactor module naming (bounds/scalar, btf/database, sanitize/spectre)
- Box large enum variants to fix large_enum_variant warnings
- Create parameter structs to reduce function arguments
- Add attach_target verification module
- Update error handling to match kernel errno conventions
- Add KBox/KVec/KString type aliases for kernel compatibility
- Implement TryAlloc and KVecExt traits for fallible allocation

Kernel integration ready:
- cargo clippy --features kernel: 0 warnings
- cargo build --features kernel: compiles successfully
- All 82 tests pass

Code quality:
- Fixed ~300+ clippy warnings
- Proper documentation for kernel modules
- Type-safe FFI bridge layer

+++ src/state/snapshot.rs
@@ -428,7 +428,7 @@ impl StateCache {
         if total == 0 {
             0
         } else {
-            ((self.hits as u64 * 100) / total as u64) as u32
+            ((self.hits * 100) / total) as u32
         }
     }
 }
